pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "receipt-processor"
version = "0.1.0"
description = "Receipt ingestion and expense export pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openpyxl>=3.1.0",
  "pypdf>=4.2.0",
  "Pillow>=10.0.0",
  "pytesseract>=0.3.10",
  "typer>=0.12.0",
  "PyYAML>=6.0.1",
  "python-dotenv>=1.0.1"
]

[project.scripts]
receipt-processor = "receipt_processor.cli:main"
receipt_processor = "receipt_processor.cli:main"
receipts_processor = "receipt_processor.cli:main"
receipts_processor_gui = "receipt_processor.gui:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
